完善收藏夹部分逻辑，修改FindID接口满足需求

+++ cmd/api/handlers/create_collection.go
@@ -20,8 +20,9 @@ func CreateCollection(c *gin.Context) {
 	createColltVar.Name = name
 
 	claim := jwt.ExtractClaims(c)
-	UserId := claim[constants.IdentityKey].(int64)
+	UserId := int64(claim[constants.IdentityKey].(float64))
 	createColltVar.UserID = UserId
+
 	req := &collectionModel.CreateColltRequest{
 		UserId: createColltVar.UserID,
 		Name:   createColltVar.Name,
